+++ src/shared/header/index.ts
@@ -1,1 +1,2 @@
-export * from './components/header';
+export * from './components/header/header';
+export * from './components/signin/signin';
